backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Warehouse
clickhouse-connect==0.7.0

# Streaming
kafka-python==2.0.2

# Redis
redis==5.0.1

# OpenAI
openai==1.3.7

# Data validation
pydantic==2.5.0

# LLM structured data extraction
instructor==0.4.8
anthropic==0.8.1

# HTTP client
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6 
python-json-logger==2.0.7 